review page and validation
