+++ src/js/projectWizard.jsx
@@ -24,7 +24,7 @@ import {
   sub_ids
 } from "./state/projectWizard";
 
-import { lengthObject, someObject, filterObject } from "./utils/sequence";
+
 
 import "../css/project-wizard.css";
 
@@ -39,227 +39,14 @@ const projectSteps = [
   {id: 'review', label: 'Review & Publish'}
 ];
 
-const successModal ={
-  id: 'success',
-  confirmText: 'Close',
-  onConfirm: ()=>{dispatch([event_ids.modal, null]);}};
 
-const errorModal = {
-  id: 'error',
-  confirmText: "Go Back",
-  onConfirm: () => {
-    dispatch([event_ids.currentStep, 'review']);
-    dispatch([event_ids.modal, null]);
-  }
-};
-
-function validateOverview ({name, description}) {
-  return(['overview', [
-    (name === "" || description === "") && "A project must contain a name and description.",
-  ]]);
-}
-
-function validateImagery ({requiresPublic, imageryId, privacyLevel}) {
-  return(['imagery', [
-    requiresPublic &&
-      `Projects with privacy level of ${privacyLevel} require at least one public imagery.`,
-      imageryId <= 0 && "Select a valid Basemap.",
-  ]]);
-}
-
-function validatePlots({
-  plotDistribution,
-  aoiFeatures,
-  plotSpacing,
-  plotSize,
-  plotFileNeeded,
-  plotFileName,
-  totalPlots,
-  plotLimit,
-  users,
-  userMethod,
-  percents,
-  qaqcMethod,
-  smes,
-  timesToReview,
-  numPlots
-  
-}) {
-  return (['plots',
-          [
-  ["random", "gridded"].includes(plotDistribution) &&
-        !aoiFeatures.length &&
-        "Please select a valid boundary.",
-      plotDistribution === "random" &&
-        !numPlots &&
-        "A number of plots is required for random plot distribution.",
-      plotDistribution === "gridded" &&
-        !plotSpacing &&
-        "A plot spacing is required for gridded plot distribution.",
-    !["shp", "geojson"].includes(plotDistribution) && (!plotSize || plotSize === 0) && "A plot size is required.",
-      plotDistribution === "csv" &&
-        plotFileNeeded &&
-        !(plotFileName || "").includes(".csv") &&
-        "A plot CSV (.csv) file is required.",
-      plotDistribution === "shp" &&
-        plotFileNeeded &&
-        !(plotFileName || "").includes(".zip") &&
-        "A plot SHP (.zip) file is required.",
-      totalPlots > plotLimit &&
-        "The plot size limit has been exceeded. Check the Plot Design section for detailed info.",
-      ["equal", "percent"].includes(userMethod) &&
-        users.length === 0 &&
-        "At least one user must be added to the plot assignment.",
-      userMethod === "percent" &&
-        percents.reduce((acc, p) => acc + p, 0) !== 100 &&
-        "The assigned plot percentages must equal 100%.",
-      userMethod === "percent" &&
-        percents.reduce((acc, p) => acc || p === 0, false) &&
-        "All plot assignment percentages must be greater than 0%.",
-      ["overlap", "sme"].includes(qaqcMethod) &&
-        percent === 0 &&
-        "The assigned Quality Control percentage must be greater than 0%.",
-      ["random", "gridded"].includes(plotDistribution) &&
-      qaqcMethod === "sme" && smes.length === 0 && "At least one user must be added as an SME.",
-      qaqcMethod === "overlap" &&
-        users.length === 1 &&
-        "At least two assigned users are required for overlap mode.",
-      qaqcMethod === "overlap" && timesToReview < 2 && "# of Reviews must be at least 2.",
-      qaqcMethod === "overlap" &&
-        timesToReview > users.length &&
-        users.length > 1 &&
-        "# of Reviews cannot be greater than the number of assigned users.",
-      userMethod !== "none" &&
-        qaqcMethod === "sme" &&
-        _.intersection(users, smes).length > 0 &&
-        "Users cannot be an Assigned User and an SME. Please remove the duplicate users.",]]);
-}
-
-function validateSamples({
-  sampleDistribution,
-  samplesPerPlot,
-  sampleResolution,
-  sampleFileName,
-  plotShape,
-  plotSize,
-  perPlotLimit,
-  totalPlots,
-  sampleLimit,
-  allowDrawnSamples,
-  sampleGeometries
-}) {
-  return ([
-    'samples',
-    [sampleDistribution === "random" &&
-     !samplesPerPlot && "A number of samples per plot is required for random sample distribution.",
-     sampleDistribution === "gridded" &&
-     !sampleResolution && "A sample spacing is required for gridded sample distribution.",
-     sampleDistribution === "csv" &&
-     !(sampleFileName || "").includes(".csv") && "A sample CSV (.csv) file is required.",
-     sampleDistribution === "shp" &&
-     !(sampleFileName || "").includes(".zip") && "A sample SHP (.zip) file is required.",
-     sampleDistribution === "gridded" &&
-     plotShape === "circle" &&
-     sampleResolution >= plotSize / Math.sqrt(2) && `You must use a sample spacing that is less than ${
-          Math.round((plotSize / Math.sqrt(2)) * 100) / 100
-        } meters.`,
-     sampleDistribution === "gridded" &&
-     plotShape === "square" &&
-     sampleResolution >= plotSize && "The sample spacing must be less than the plot width.",
-     (samplesPerPlot > perPlotLimit || totalPlots * samplesPerPlot > sampleLimit) &&
-     "The sample size limit has been exceeded. Check the Sample Design section for detailed info.",
-     allowDrawnSamples && !Object.values(sampleGeometries).some((g) => g) &&
-     "At least one geometry type must be enabled.",]]);
-}
-
-function validateQuestions({
-  surveyQuestions
-}) {
-  function allAnswersHidden (answers) {
-    const hiddenAnswers = filterObject(answers, ([_id, ans]) => ans.hide);
-    return lengthObject(answers) === lengthObject(hiddenAnswers);}
-
-  return ([
-    'questions',
-    [lengthObject(surveyQuestions) === 0 && "A survey must include at least one question.",
-     someObject(surveyQuestions, ([_id, sq]) => (lengthObject(sq.answers) === 0 || allAnswersHidden(sq.answers))) &&
-     "All survey questions must contain at least one (unhidden) answer.",]
-  ]);
-}
-
-function validateWizard () {
-  /* validates the existing draft project.
-     returns a boolean.
-     side effects: sends api request to create project,
-     or adds errors to state.
-     may even send a request, if superficially valid, and still add errors to state, in case of server error
-  */
-  const form = {name : useSubscription([sub_ids.overview.projectName]),
-	        description : useSubscription([sub_ids.overview.projectDescription]),
-
-	        privacyLevel : useSubscription([sub_ids.overview.visibility]),
-
-                //TODO: what are these validation values?
-                //imageryId: is this really just one imagery?
-	        imageryId : useSubscription([sub_ids.imagery.imagery])[0],
-	        institutionImagery : [],
-	        projectImageryList : [],
-	        requiresPublic : false,
-
-                //TODO: what is this value? not validated against
-	        projectId : -1,
-                
-	        aoiFeatures : useSubscription([sub_ids.boundary.aoiFeatures]),
-	        plotDistribution : useSubscription([sub_ids.plots.plotDistribution]),
-	        numPlots : useSubscription([sub_ids.plots.numPlots]),
-	        plotSpacing: useSubscription([sub_ids.plots.plotSpacing]),
-	        plotSize : useSubscription([sub_ids.plots.plotSize]),
-	        plotFileName: usEsubscription([sub_ids.plots.plotFileName]),
-
-                //TODO: are we using these values for validation?
-	        useTemplatePlots : false,
-                //originalProject ,
-                
-	        designSettings: useSubscription([sub_ids.plots.designSettings]),
-	        totalPlots : useSubscription([sub_ids.plots.totalPlots]),
-
-                //TODO: how is plotFileNeeded used/stored? required for validation
-	        plotFileNeeded : false,
-
-                
-	        allowDrawnSamples : useSubscription([sub_ids.samples.allowDrawnSamples]),
-	        samplesPerPlot : useSubscription([sub_ids.samples.samplesPerPlot]),
-	 
-	        plotShape : useSubscription([sub_ids.plots.plotShape]),
-	        sampleDistribution : useSubscription([sub_ids.samples.sampleDistribution]),
-	        sampleFileName : useSubscription([sub_ids.samples.sampleFileName]),
-	        sampleResolution : useSubscription([sub_ids.samples.sampleResolution]),
-
-	        surveyQuestions : useSubscription([sub_ids.questions.questions])};
-
-  const errors = [
-    validateOverview(form),
-    validateImagery(form),
-    validatePlots(form),
-    validateSamples(form),
-    validateQuestions(form)
-  ].filter(([name, e])=>e.length);
-  
-  return (errors.length ? false : errors);
-}
 
 const publishModal = {
   title: 'Ready to publish this project?',
   id: 'review',
   closeText: "Cancel",
   confirmText: "Publish Project",
-  onConfirm: ()=>{
-    let errors = validateWizard();
-    errors ?
-      dispatch ([event_ids.modal, errorModal]) //dispatch an event that adds errors to state and then sends errors
-      : dispatch([event_ids.modal, successModal]); //dispatch an event that sends the api request and eventually succesModal
-      
-  }
+  onConfirm: ()=>{dispatch ([event_ids.validate]); }
 };
 
 const exitModal = {
@@ -276,12 +63,11 @@ const exitModal = {
 
 };
 
-const NavButtons = () => {
+function NavButtons  () {
   const currentStep = useSubscription([sub_ids.currentStep]);
   
   const  continueHandler = () => {
     const currentIdx = projectSteps.findIndex(({id})=> (id == currentStep));
-    
     currentIdx + 1 < projectSteps.length
       ? dispatch([event_ids.currentStep, projectSteps[currentIdx + 1].id])
       : dispatch([event_ids.modal, publishModal]);
@@ -318,8 +104,8 @@ const ProjectWizardNavigator = () => {
             <div
               key={id}
               style={{fontWeight: currentStep === id ? 'bold' : 'normal',
-                display: 'inline-flex',
-                cursor: 'pointer'}}
+                      display: 'inline-flex',
+                      cursor: 'pointer'}}
               onClick={
                 () => dispatch([event_ids.currentStep, id])
               }
@@ -337,11 +123,11 @@ const ProjectWizardNavigator = () => {
 function NewProjectModal () {
   const newProjectOptions = {
     newProject: ['Create a new project',
-      'Generate a new project from scratch by customizing all steps.'],
+                 'Generate a new project from scratch by customizing all steps.'],
     templateProject: ['Select from an existing template',
-      'Select a template and prefill all the steps. You can edit and customize it.'],
+                      'Select a template and prefill all the steps. You can edit and customize it.'],
     importProject: ['Import Collect Earth Project',
-      'Need Description']};
+                    'Need Description']};
   const projectSource = useSubscription([sub_ids.projectSource]);
 
   return (
@@ -351,8 +137,8 @@ function NewProjectModal () {
         return (
           <div
             className={projectSource === id ?
-              "radio-selected-button"
-              : "radio-selection-button"}
+                       "radio-selected-button"
+                       : "radio-selection-button"}
             key={id}
             onClick={()=> {
               dispatch([event_ids.projectSource, id]);
@@ -362,7 +148,7 @@ function NewProjectModal () {
             >{projectSource === id
               ? <SvgIcon icon="radioChecked" size="1.2rem" />                            
               : <SvgIcon icon="radio" size="1.2rem"
-                         className="radio-button-unchecked"/> }
+                                          className="radio-button-unchecked"/> }
               {"    "}
               { title } </p>
             <label
@@ -409,6 +195,7 @@ function ErrorModal () {
   );
 };
 
+
 function ExitModal () {
   return (
     <p > Are you sure you want to exit “Create New Scenario”? Saved steps will be kept in draft form, any steps you haven’t saved will be lost.</p>
@@ -419,6 +206,7 @@ const ProjectWizardModal = () => {
   // this is the container for any modal related to this page. based on state, this actually renders modals as they are explicitly defined above., provided through "children" value of modal map"
   const projectSource = useSubscription([sub_ids.projectSource]);
   const modal = useSubscription([sub_ids.modal]);
+
   
   const children = () => {
     switch (modal.id) {
@@ -460,6 +248,8 @@ const ProjectWizard = ({userId, userName, version, institutionId}) => {
   
   const currentStep = useSubscription([sub_ids.currentStep]);
   const modal = useSubscription([sub_ids.modal]);
+
+
   
   // -------------------
   // HANDLERS
@@ -508,7 +298,7 @@ const ProjectWizard = ({userId, userName, version, institutionId}) => {
     case 'boundary'   : return <BoundaryStep />;
     case 'plots'      : return <PlotStep />;
     case 'samples'    : return <SampleStep />;
-    case 'questions'  : return <SurveyQuestionsStep />;
+    case 'questions'  : return <SurveyQuestionsStep/>;
     case 'rules'      : return <RulesStep />;
     case 'review'     : return <ReviewStep imageryList={availableImagery}/>;
     default           : return <div style={{padding: "20px"}}>Step {currentStep} coming soon</div>;
@@ -522,15 +312,15 @@ const ProjectWizard = ({userId, userName, version, institutionId}) => {
         <BreadCrumbs
           crumbs={[
             {display: "Institution",
-              id: "institution",
-              query: ["institution", institutionId],
-              onClick:()=>{window.location.assign(`/review-institution?institutionId=${institutionId}`);
-              }},
+             id: "institution",
+             query: ["institution", institutionId],
+             onClick:()=>{window.location.assign(`/review-institution?institutionId=${institutionId}`);
+                         }},
             {display: "Add a New Project",
-              id: "projectWizard",
-              query: ["project", "newProject"],
-              onClick:()=>{window.location.assign(`/project-wizard?institutionId=${institutionId}`);
-              }}]}
+             id: "projectWizard",
+             query: ["project", "newProject"],
+             onClick:()=>{window.location.assign(`/project-wizard?institutionId=${institutionId}`);
+                         }}]}
         />
         <ProjectWizardNavigator/>
         <div className="wizard-step-body" >
